+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPFA4WRHG9PJ40P25FM5BEA"
+        "01KNPN6KHCK720AE7T8H84VA3P"
     )
     r"""The unique identifier of the data source"""